This change reworks the org.marketcetera.marketdata.MarketDataRequest object to address the requirements for this release as well as predicted changes for the future.  The original implementation made a number of assumptions in the absence of real requirements.  For the most part, the structure of the request mirrored FIX Market Data Requests.  The current implementation goes back to square one and creates a representation that is better-suited to the requirements that now exist.  The current implementation is backwards-compatible from the point-of-view of strategy.
The changes are largely centered in MarketDataRequest and its test.  These files were essentially rewritten, so the diffs are likely to be complicated.  It may be better to simply take MarketDataRequest and MarketDataRequestTest as new.

One particularly interesting change is in MarketDataRequest.DateUtils.  This class provides support for parsing and emitting dates for the String representation of market data requests needed for Strategy Agent.  The requirement for OHLC data (Open High Low Close) includes the specification of a date.  Parsing dates is a real can-of-worms (or warm bucket-of-poo depending on your point-of-view) because of L10n issues.  I decided to go with an ISO standard for date specification that at least will be equally uncomfortable for everybody.  The ISO is 8601 and is actually widely used in software, so it shouldn't be too problematic.  I included utilties to translate from Java Date to ISO and vice-versa to try to reduce the complexity to the user but, for SA requests, there's no getting around it.  This method allows specification of dates without punctuation (which can differ per locale) and in the same order across locales (year-month-day regardless).  I would be in favor of this moving to Utils, but I'll let T2 decide that.

Changes were also made to all four market data adapters to react to the new MarketDataRequest format.  All four adapters can now respond to the new requests and their service level remains the same (could handle TOP_OF_BOOK requests before can handle them now).  There is a complimentary review on the private side.

All unit tests pass, including Photon tests.  Manual testing passes, too.  The unit tests for the market data apadater message translation are largely new and represent part of an effort to beef up difficult-to-test market data adapters.